refactor(@dpc-sdp/ripple-ui-forms): :art: move form input definitions outside of plugin
Renames to components PascalCase
