+++ .vscode/settings.json
@@ -19,6 +19,7 @@
   "cSpell.words": [
     "colour",
     "formkit",
+    "maxlength",
     "nuxt",
     "pathe",
     "scripthost",
