+++ packages/ripple-ui-forms/src/components/RplFormInput/RplFormInput.stories.mdx
@@ -4,7 +4,7 @@ import {
   Story,
   ArgsTable
 } from '@storybook/addon-docs'
-import RplFormInput from './input.vue'
+import RplFormInput from './RplFormInput.vue'
 
 export const SingleTemplate = (args) => ({
   components: { RplFormInput },
