--- packages/ripple-ui-forms/src/components/select/select.stories.mdx
@@ -1,33 +1,0 @@
-import {
-  Canvas,
-  Meta,
-  Story,
-  ArgsTable
-} from '@storybook/addon-docs'
-import RplFormSelect from './select.vue'
-
-export const SingleTemplate = (args) => ({
-  components: { RplFormSelect },
-  setup() {
-    return { args }
-  },
-  template: '<RplTextInput v-bind="args" />'
-})
-
-<Meta title='Forms/Text' component={RplFormSelect} />
-
-# Forms
-
-<ArgsTable of={RplFormSelect} />
-
-<Canvas>
-  <Story
-    name='Default'
-    args={{
-      value: 'test'
-    }}
-  >
-    {SingleTemplate.bind()}
-  </Story>
-</Canvas>
-
