+++ packages/ripple-ui-forms/src/index.ts
@@ -1,4 +1,4 @@
-export { default as RplForm } from './components/form/form.vue'
-export { default as RplFormElement } from './components/element/element.vue'
+export { default as RplForm } from './components/RplForm/RplForm.vue'
+export { default as RplFormElement } from './components/RplFormElement/RplFormElement.vue'
 
 export { default as registerRplFormPlugin } from './register.js'
