+++ packages/ripple-ui-forms/src/inputs/index.ts
@@ -1,7 +1,4 @@
-export { checkbox } from './checkbox'
-export { text as date } from './date'
-export { radio } from './radio'
-export { select } from './select'
-export { text as tel } from './tel'
 export { text } from './text'
-export { button as submit } from './submit'
+export { email } from './email'
+export { tel } from './tel'
+export { submit } from './submit'
