+++ packages/ripple-ui-forms/src/inputs/text.ts
@@ -1,49 +1,26 @@
-import { markRaw } from 'vue'
 import { FormKitTypeDefinition } from '@formkit/core'
 import {
-  outer,
-  inner,
-  wrapper,
-  label,
-  help,
-  messages,
-  message,
-  icon,
-  prefix,
-  suffix,
-  createSection
-} from '@formkit/inputs'
-
-import RplFormText from './../components/text/text.vue'
+  createRplFormInput,
+  defaultRplFormInputProps,
+  inputLibrary
+} from './input-utils'
 
 /**
- * Input definition for a text.
+ * Input definition for Ripple text input.
  * @public
  */
 export const text: FormKitTypeDefinition = {
   /**
    * The actual schema of the input, or a function that returns the schema.
    */
-  schema: outer(
-    wrapper(
-      label('$label'),
-      help('$help'),
-      inner(
-        icon('prefix', 'label'),
-        prefix(),
-        createSection('input', () => ({
-          $cmp: 'rplFormText',
-          props: {
-            context: '$node.context'
-          }
-        }))(),
-        suffix(),
-        icon('suffix')
-      )
-    ),
-    messages(message('$message.value'))
-  ),
-  library: { rplFormText: markRaw(RplFormText) },
+  schema: createRplFormInput({
+    $cmp: 'RplFormInput',
+    props: {
+      ...defaultRplFormInputProps,
+      type: 'text'
+    }
+  }),
+  library: inputLibrary,
   /**
    * The type of node, can be a list, group, or input.
    */
@@ -56,7 +33,7 @@ export const text: FormKitTypeDefinition = {
   /**
    * An array of extra props to accept for this input.
    */
-  props: [],
+  props: ['placeholder', 'validationMeta'],
   /**
    * Forces node.props.type to be this explicit value.
    */
