+++ packages/ripple-ui-forms/src/register.ts
@@ -1,9 +1,5 @@
 import { plugin, defaultConfig } from '@formkit/vue'
-import defaultRplFormConfig from './formkit.config'
 
-export default function registerRplFormPlugin(
-  vueApp,
-  config = defaultRplFormConfig
-) {
-  vueApp.use(plugin, defaultConfig(config))
+export default function registerRplFormPlugin(vueApp, config = defaultConfig) {
+  vueApp.use(plugin, config)
 }
